+++ Source/Steam Library Manager/Forms/Main.xaml.cs
@@ -2,6 +2,7 @@
 using NLog;
 using NLog.Targets.Wrappers;
 using System;
+using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Linq;
 using System.Windows;
@@ -16,7 +17,54 @@ namespace Steam_Library_Manager
     public partial class Main
     {
         public static Main FormAccessor;
-        public Framework.AsyncObservableCollection<string> TaskManager_Logs = new Framework.AsyncObservableCollection<string>();
+        private ObservableCollection<string> TMLogs { get; } = new ObservableCollection<string>();
+
+        public IProgress<string> TaskManager_Logs = new Progress<string>(log => FormAccessor.TMLogs.Add(log));
+
+        public IProgress<Definitions.Library> LibraryChange = new Progress<Definitions.Library>(Library =>
+        {
+            if (Library == null)
+                return;
+
+            string Search = (Properties.Settings.Default.includeSearchResults) ? Properties.Settings.Default.SearchText : null;
+            if (Definitions.List.Libraries.Count(x => x == Library) == 0 || !Library.DirectoryInfo.Exists)
+            {
+                FormAccessor.AppView.AppPanel.ItemsSource = null;
+                return;
+            }
+
+            Func<dynamic, object> Sort = Functions.SLM.Settings.GetSortingMethod(Library);
+
+            switch (Library.Type)
+            {
+                case Definitions.Enums.LibraryType.Steam:
+                case Definitions.Enums.LibraryType.SLM:
+                    FormAccessor.AppView.AppPanel.ItemsSource = (Properties.Settings.Default.defaultGameSortingMethod == "sizeOnDisk" || Properties.Settings.Default.defaultGameSortingMethod == "LastUpdated" || Properties.Settings.Default.defaultGameSortingMethod == "LastPlayed") ?
+                        ((string.IsNullOrEmpty(Search)) ?
+                        Library.Steam.Apps.OrderByDescending(Sort).ToList() : Library.Steam.Apps.Where(
+                            y => y.AppName.IndexOf(Search, StringComparison.InvariantCultureIgnoreCase) >= 0 || y.AppID.ToString().Contains(Search) // Search by app ID
+                        ).OrderByDescending(Sort).ToList()
+                        ) :
+                        ((string.IsNullOrEmpty(Search)) ? Library.Steam.Apps.OrderBy(Sort).ToList() : Library.Steam.Apps.Where(
+                        y => y.AppName.IndexOf(Search, StringComparison.InvariantCultureIgnoreCase) >= 0 || y.AppID.ToString().Contains(Search) // Search by app ID
+                        ).OrderBy(Sort).ToList()
+                        );
+                    break;
+
+                case Definitions.Enums.LibraryType.Origin:
+                    FormAccessor.AppView.AppPanel.ItemsSource = (Properties.Settings.Default.defaultGameSortingMethod == "sizeOnDisk" || Properties.Settings.Default.defaultGameSortingMethod == "LastUpdated") ?
+                        ((string.IsNullOrEmpty(Search)) ?
+                        Library.Origin.Apps.OrderByDescending(Sort).ToList() : Library.Origin.Apps.Where(
+                            y => y.AppName.IndexOf(Search, StringComparison.InvariantCultureIgnoreCase) >= 0 || y.AppID.ToString().Contains(Search) // Search by app ID
+                        ).OrderByDescending(Sort).ToList()
+                        ) :
+                        ((string.IsNullOrEmpty(Search)) ? Library.Origin.Apps.OrderBy(Sort).ToList() : Library.Origin.Apps.Where(
+                        y => y.AppName.IndexOf(Search, StringComparison.InvariantCultureIgnoreCase) >= 0 || y.AppID.ToString().Contains(Search) // Search by app ID
+                        ).OrderBy(Sort).ToList()
+                        );
+                    break;
+            }
+        });
 
         public Main()
         {
@@ -51,7 +99,7 @@ namespace Steam_Library_Manager
 
                 TaskManagerView.TaskPanel.ItemsSource = Framework.TaskManager.TaskList;
                 TaskManagerView.TaskManagerInformation.DataContext = Framework.TaskManager.TMInfo;
-                TaskManagerView.TaskManager_LogsView.ItemsSource = TaskManager_Logs;
+                TaskManagerView.TaskManager_LogsView.ItemsSource = TMLogs;
 
                 LibraryCleanerView.LibraryCleaner.ItemsSource = Definitions.List.LCItems;