+++ Source/Steam Library Manager/Framework/TaskManager.cs
@@ -45,10 +45,10 @@ namespace Steam_Library_Manager.Framework
                     {
                         if (CurrentTask.RemoveOldFiles && CurrentTask.TaskType != Definitions.Enums.TaskType.Delete)
                         {
-                            Main.FormAccessor.TaskManager_Logs.Add(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFiles)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.SteamApp.AppName }));
+                            Main.FormAccessor.TaskManager_Logs.Report(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFiles)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.SteamApp.AppName }));
                             await CurrentTask.SteamApp.DeleteFilesAsync(CurrentTask);
                             CurrentTask.SteamApp.Library.Steam.Apps.Remove(CurrentTask.SteamApp);
-                            Main.FormAccessor.TaskManager_Logs.Add(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFilesCompleted)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.SteamApp.AppName }));
+                            Main.FormAccessor.TaskManager_Logs.Report(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFilesCompleted)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.SteamApp.AppName }));
                         }
 
                         if (CurrentTask.TargetLibrary?.Type == Definitions.Enums.LibraryType.Steam)
@@ -88,12 +88,12 @@ namespace Steam_Library_Manager.Framework
                     {
                         if (CurrentTask.RemoveOldFiles && CurrentTask.TaskType != Definitions.Enums.TaskType.Delete)
                         {
-                            Main.FormAccessor.TaskManager_Logs.Add(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFiles)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.OriginApp.AppName }));
+                            Main.FormAccessor.TaskManager_Logs.Report(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFiles)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.OriginApp.AppName }));
 
                             CurrentTask.OriginApp.DeleteFiles(CurrentTask);
                             CurrentTask.OriginApp.Library.Origin.Apps.Remove(CurrentTask.OriginApp);
 
-                            Main.FormAccessor.TaskManager_Logs.Add(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFilesCompleted)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.OriginApp.AppName }));
+                            Main.FormAccessor.TaskManager_Logs.Report(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_RemoveOldFilesCompleted)), new { CurrentTime = DateTime.Now, AppName = CurrentTask.OriginApp.AppName }));
                         }
 
                         CurrentTask.TaskStatusInfo = Functions.SLM.Translate(nameof(Properties.Resources.TaskStatus_Completed));
@@ -153,7 +153,7 @@ namespace Steam_Library_Manager.Framework
         {
             if (!Status && !Paused)
             {
-                Main.FormAccessor.TaskManager_Logs.Add(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_Active)), new { CurrentTime = DateTime.Now }));
+                Main.FormAccessor.TaskManager_Logs.Report(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_Active)), new { CurrentTime = DateTime.Now }));
                 CancellationToken = new CancellationTokenSource();
                 Status = true;
 
@@ -174,7 +174,7 @@ namespace Steam_Library_Manager.Framework
                     catch (OperationCanceledException)
                     {
                         Stop();
-                        Main.FormAccessor.TaskManager_Logs.Add(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_Stopped)), new { CurrentTime = DateTime.Now }));
+                        Main.FormAccessor.TaskManager_Logs.Report(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_Stopped)), new { CurrentTime = DateTime.Now }));
                     }
                     catch (Exception ex)
                     {
@@ -225,7 +225,7 @@ namespace Steam_Library_Manager.Framework
                     Paused = true;
                     ActiveTask.mre.Reset();
 
-                    Main.FormAccessor.TaskManager_Logs.Add(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_Paused)), new { CurrentTime = DateTime.Now }));
+                    Main.FormAccessor.TaskManager_Logs.Report(StringFormat.Format(Functions.SLM.Translate(nameof(Properties.Resources.TaskManager_Paused)), new { CurrentTime = DateTime.Now }));
                 }
             }
             catch (Exception ex)
@@ -258,6 +258,7 @@ namespace Steam_Library_Manager.Framework
                     CancellationToken.Cancel();
                     IsRestartRequired = false;
                     ActiveTask?.mre?.Set();
+                    TmInfoUpdate();
                 }
             }
             catch (Exception ex)