+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+## [1.5.1.1] - 2019-04-17
+
+### Added
+
+- Task Manager Logs Auto Scroll on/off switch
+
+### Fixed
+
+- (Task Manager) Tasked item crash with: System.IndexOutOfRangeException: Index was outside the bounds of the array.
+- Origin game header images are not downloading: RuntimeBinderException
+
 ## [1.5.1.0] - 2019-04-06
 
 ### Added